+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Bug Fixes
 - Fix errors in loading a scaled / transformed raster image
 - Limit image extent validation to the bbox that will actually be engraved (skipping eg overhanging whitespace)
+- Fix transparency in jpegs going black
 
 ### Development
 -